[FIX] LocalDynamodb config 추가
